Add main.rb
