+++ app.rb
@@ -152,38 +152,3 @@ class App
     person_obj[0]
   end
 end
-
-def main
-  item = 0
-  app = App.new
-  while item != 7
-    puts '1 - List all books'
-    puts '2 - List all people'
-    puts '3 - Create a person'
-    puts '4 - Create a book'
-    puts '5 - Create a rental'
-    puts '6 - List all rentals for a given person ID'
-    puts '7 - Exit'
-    item = gets.chomp.to_i
-    choose(item, app)
-  end
-end
-
-def choose(item, app)
-  case item
-  when 1
-    app.list_all_books
-  when 2
-    app.list_all_people
-  when 3
-    app.create_person
-  when 4
-    app.create_book
-  when 5
-    app.create_rental
-  when 6
-    app.list_rentals
-  end
-end
-
-main
